+++ packages/terra-button-group/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports
 
 3.15.0 - (June 28, 2019)
 ------------------